fix()[develop]: fix errors in several classes. set paginated in base repository.
