--- backend/src/auth/application/Guards/JwtAuthGuard.ts
@@ -1,7 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { AuthGuard } from "@nestjs/passport";
-
-@Injectable()
-export class JwtAuthGuard extends AuthGuard('jwt'){
-
-}
