+++ backend/src/auth/application/dtos/returnLogin.dto.ts
@@ -1,0 +1,3 @@
+export type ReturnLoginDto = {
+  access_token: string
+}
