--- backend/src/auth/application/dtos/returnLoginDto.ts
@@ -1,3 +1,0 @@
-export class ReturnLoginDto {
-    acces_token: string
-}
