+++ backend/src/auth/application/guards/jwtAuthGuard.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
+
+@Injectable()
+export class JwtAuthGuard extends AuthGuard('jwt'){
+
+}
