+++ backend/src/auth/application/guards/localAuthGuard.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
+
+@Injectable()
+export class LocalAuthGuard extends AuthGuard('local'){
+  
+}
