+++ backend/src/auth/application/strategies/localStrategy.ts
@@ -1,8 +1,9 @@
-import { Injectable, UnauthorizedException } from "@nestjs/common";
-import { PassportStrategy } from "@nestjs/passport"
-import { Strategy } from 'passport-local'
-import { User } from "src/user/domain/entities/user.entity";
-import { ValidateUserUseCase } from "../useCase/auth.validate.use-case";
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { PassportStrategy } from '@nestjs/passport';
+import { Strategy } from 'passport-local';
+import { User } from 'src/user/domain/entities/user.entity';
+import { ValidateUserUseCase } from '../useCase/auth.validate.use-case';
+
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
     constructor(private readonly validateUserUseCase: ValidateUserUseCase) {
