+++ backend/src/user/application/dtos/user.create.dto.ts
@@ -1,10 +1,10 @@
-import { EnumPermits } from "src/shared/domain/enum.permits";
-import { EnumStatus } from "src/user/domain/enums/enum.status";
+import { EnumPermits } from 'src/shared/domain/enum.permits';
+import { UserStatus } from 'src/user/domain/enums/user.status';
 
 export type UserCreateDto = {
     username: string;
     roles: EnumPermits[]
     password: string
     email: string
-    status: EnumStatus
+    status: UserStatus
 }
