+++ backend/src/user/application/dtos/user.dto.ts
@@ -1,10 +1,10 @@
 import { EnumPermits } from 'src/shared/domain/enum.permits';
 import { BaseDto } from '../../../shared/core/BaseDto';
-import { EnumStatus } from '../../domain/enums/enum.status';
+import { UserStatus } from '../../domain/enums/user.status';
 
 export type UserDto = BaseDto & {
   username: string;
   roles: EnumPermits[];
   email: string;
-  status: EnumStatus;
+  status: UserStatus;
 }
