+++ backend/src/user/application/useCases/user.create.use-case.ts
@@ -2,12 +2,10 @@ import { Either, left, right } from 'src/shared/core/Either';
 import { AppError } from '../../../shared/core/errors/AppError';
 import { Result } from '../../../shared/core/Result';
 import { IUseCase } from '../../../shared/core/interfaces/IUseCase';
-import { Injectable, Logger, Optional } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { User } from 'src/user/domain/entities/user.entity';
 import { UserCreateDto } from '../dtos/user.create.dto';
 import { UserRepository } from 'src/user/infra/repositories/user.repository';
-import { EnumStatus } from 'src/user/domain/enums/enum.status';
-import { EnumPermits } from 'src/shared/domain/enum.permits';
 
 export type CreateUserUseCaseResponse = Either<AppError.UnexpectedErrorResult<User>
     | AppError.ValidationErrorResult<User>,
